Add s3 root query
